Updated ML model to include entire alphabet
NOTE: to properly run model with app.py, change name of keypoint_classifier1.tflite to keypoint_classifier.tflite and make sure that keypoint_classifier_label.csv is replaced with the one in this commit. Both of these files should be in the model folder.

Also included keypoint.csv which has all the merged training data. You DO NOT need to download keypoint.csv for the model to run. It is only there for reference or if you want to retrain the model. PLEASE DO NOT add keypoint.csv to the Android Studio project.